recipes/natural_language_processing/graph_rag/app/requirements.txt
langchain-openai==0.1.7
langchain==0.1.20
sentence-transformers==2.7.0
streamlit==1.34.0
neo4j==5.23.0
neo4j-graphrag==1.0.0
pypdf==4.3.1
langchain_experimental
python-dotenv
